Progress with monkey swing collision

+++ TR5Main/Game/Lara/lara_helpers.h
@@ -9,7 +9,7 @@ struct COLL_INFO;
 // For State Control & Collision
 // -----------------------------
 
-void DoLaraLean(ITEM_INFO* item, COLL_INFO* coll, int maxAngle, short rate);
+void DoLaraLean(ITEM_INFO* item, COLL_INFO* coll, short maxAngle, short rate);
 void DoLaraStep(ITEM_INFO* item, COLL_INFO* coll);
 void DoLaraCrawlVault(ITEM_INFO* item, COLL_INFO* coll);
 void DoLaraCrawlToHangSnap(ITEM_INFO* item, COLL_INFO* coll);
